Pull wire: price is i32 cents end-to-end (kills a silent width drift)
The specta-side repo audit spotted PullIngredient.price disagreeing across
the wire: Option<i64> on the server, Option<i32> on the desktop mirror. The
truth follows the write path - SaveIngredientInput.price (the ONLY way a
price is written) is i32 cents, so every stored value is i32-range and the
server's i64 was the outlier, not the desktop's i32. Narrowed the server
pull field to the real contract; desktop, core, and TS (number) all agree.

Not a live fault (no price can exceed i32 through the DAL) but a landmine:
the exact silent-numeric-width bug class as specta #481/#508, sitting in
our own sync path - in a repo whose TS mirrors have a drift gate while the
Rust-to-Rust mirrors have none. The durable fix is the declared
vegify-client-rs direction (shared/generated wire types); this closes the
hole today.

A full field-by-field re-read of all five Pull* pairs found no other width
or optionality mismatch (the String->Visibility enum refinement and the
serde(default) skew tolerances on the desktop are deliberate).

+++ services/api/src/content.rs
@@ -47,7 +47,10 @@ pub struct PullIngredient {
     pub visibility: String,
     pub name: String,
     pub description: Option<String>,
-    pub price: Option<i64>,
+    /// Cents (USD). i32 END-TO-END: SaveIngredientInput (the only write path) is i32, the desktop
+    /// mirror is i32, the TS side is number — an i64 here was silent width drift in the wire contract
+    /// (found by the specta-side repo audit; the class of bug a generated client would make impossible).
+    pub price: Option<i32>,
     pub calories_per_100g: Option<f64>,
     pub serving_grams: Option<f64>,
     pub package_grams: Option<f64>,
